+++ src/math.cpp
@@ -9,7 +9,7 @@
 
 // order of operation matters! M1 * M2 represents the left side multiplication of M2 to M1 and M2 * M1 the right side multiplication of M2 to M1.
 
-std::vector<std::vector<float>> operator*(std::vector<std::vector<float>>& firstMatrix, std::vector<std::vector<float>>& secondMatrix) // Matrix - Matrix multiplication - overloaded operator    
+std::vector<std::vector<double>> operator*(std::vector<std::vector<double>>& firstMatrix, std::vector<std::vector<double>>& secondMatrix) // Matrix - Matrix multiplication - overloaded operator    
 {
     size_t m1, n1;
     size_t m2, n2;
@@ -23,7 +23,7 @@ std::vector<std::vector<float>> operator*(std::vector<std::vector<float>>& first
     if (n1 != m2)
         throw 1; // Incompatible dimensions
 
-    std::vector<std::vector<float>> result(m1, std::vector<float>(n2));
+    std::vector<std::vector<double>> result(m1, std::vector<double>(n2));
 
     for (size_t i = 0; i < m1; ++i) // row selection
     {
@@ -39,7 +39,7 @@ std::vector<std::vector<float>> operator*(std::vector<std::vector<float>>& first
     return result;
 }
 
-std::vector<std::vector<float>> multiplyMatrices(std::vector<std::vector<float>>& firstMatrix, std::vector<std::vector<float>>& secondMatrix) // Matrix - Matrix multiplication
+std::vector<std::vector<double>> multiplyMatrices(std::vector<std::vector<double>>& firstMatrix, std::vector<std::vector<double>>& secondMatrix) // Matrix - Matrix multiplication
 {
     size_t m1, n1;
     size_t m2, n2;
@@ -53,7 +53,7 @@ std::vector<std::vector<float>> multiplyMatrices(std::vector<std::vector<float>>
     if (n1 != m2)
         throw 1; // Incompatible dimensions
 
-    std::vector<std::vector<float>> result(m1, std::vector<float>(n2));
+    std::vector<std::vector<double>> result(m1, std::vector<double>(n2));
 
     for (size_t i = 0; i < m1; ++i) // row selection
     {
@@ -69,11 +69,11 @@ std::vector<std::vector<float>> multiplyMatrices(std::vector<std::vector<float>>
     return result;
 }
 
-std::vector<float> operator*(std::vector<std::vector<float>>& Matrix, std::vector<float>& Vector) // Matrix - Vector multiplication - overloaded operator
+std::vector<double> operator*(std::vector<std::vector<double>>& Matrix, std::vector<double>& Vector) // Matrix - Vector multiplication - overloaded operator
 {
     size_t m1, n1;
     size_t m2, n2;
-    float epsilon = 1e-6;
+    double epsilon = 1e-6;
 
     m1 = Matrix.size(); // rows
     n1 = Matrix[0].size(); // columns
@@ -83,7 +83,7 @@ std::vector<float> operator*(std::vector<std::vector<float>>& Matrix, std::vecto
     if (n1 != m2)
         throw 1; // Incompatible dimensions
 
-    std::vector<float> result(m1);
+    std::vector<double> result(m1);
 
     for (size_t i = 0; i < m1; ++i)
     {
@@ -98,11 +98,11 @@ std::vector<float> operator*(std::vector<std::vector<float>>& Matrix, std::vecto
     return result;
 }
 
-std::vector<float> multiplyMatrixVector(const std::vector<std::vector<float>>& Matrix, const std::vector<float>& Vector) // Matrix - Vector multiplication
+std::vector<double> multiplyMatrixVector(const std::vector<std::vector<double>>& Matrix, const std::vector<double>& Vector) // Matrix - Vector multiplication
 {
     size_t m1, n1;
     size_t m2, n2;
-    float epsilon = 1e-6;
+    double epsilon = 1e-6;
 
     m1 = Matrix.size(); // rows
     n1 = Matrix[0].size(); // columns
@@ -112,7 +112,7 @@ std::vector<float> multiplyMatrixVector(const std::vector<std::vector<float>>& M
     if (n1 != m2)
         throw 1; // Incompatible dimensions
 
-    std::vector<float> result(m1);
+    std::vector<double> result(m1);
 
     for (size_t i = 0; i < m1; ++i)
     {
@@ -136,16 +136,16 @@ bool randBool()
     return randomBool;
 }
 
-float randProb()
+double randProb()
 {
-    float randomProbability;
+    double randomProbability;
     std::random_device rd;
-    std::uniform_real_distribution<float> dist(0, 1);
+    std::uniform_real_distribution<double> dist(0, 1);
     randomProbability = dist(rd);
     return randomProbability;
 }
 
-int preSort(std::vector<std::vector<float>>& orderedVector) // puts all non-zero values to the front of the vector (zero values are irrelevant for probability)
+int preSort(std::vector<std::vector<double>>& orderedVector) // puts all non-zero values to the front of the vector (zero values are irrelevant for probability)
 {
     size_t n = orderedVector[0].size();
     int swaps = 0;
@@ -162,7 +162,7 @@ int preSort(std::vector<std::vector<float>>& orderedVector) // puts all non-zero
     return swaps - 1; // acts as indicator where the zeros start
 }
 
-void quickSort(std::vector<std::vector<float>>& orderedVector, int startingPoint, int endingPoint)
+void quickSort(std::vector<std::vector<double>>& orderedVector, int startingPoint, int endingPoint)
 {
     if (startingPoint >= endingPoint)
         return;
@@ -173,9 +173,9 @@ void quickSort(std::vector<std::vector<float>>& orderedVector, int startingPoint
 
 }
 
-int partition(std::vector<std::vector<float>>& orderedVector, int startingPoint, int endingPoint)
+int partition(std::vector<std::vector<double>>& orderedVector, int startingPoint, int endingPoint)
 {
-    float pivot = orderedVector[1][startingPoint];
+    double pivot = orderedVector[1][startingPoint];
     int i = startingPoint;
 
     for (int j = startingPoint + 1; j <= endingPoint; ++j)
@@ -193,18 +193,18 @@ int partition(std::vector<std::vector<float>>& orderedVector, int startingPoint,
     return i;
 }
 
-std::vector<float> mtMm(std::vector<std::vector<float>>& Matrix, std::vector<float>& Vector) //multi-thread Matrix multiplication
+std::vector<double> mtMm(std::vector<std::vector<double>>& Matrix, std::vector<double>& Vector) //multi-thread Matrix multiplication
 {
     size_t n = Matrix.size();
     size_t w1 = floor(n / 2);
 
-    std::vector<std::vector<float>> M1(Matrix.begin(), Matrix.begin() + w1);
-    std::vector<std::vector<float>> M2(Matrix.begin() + w1, Matrix.end());
-    auto multiplyMatrixVectorPtr = static_cast<std::vector<float>(*)(const std::vector<std::vector<float>>&, const std::vector<float>&)>(&multiplyMatrixVector);
+    std::vector<std::vector<double>> M1(Matrix.begin(), Matrix.begin() + w1);
+    std::vector<std::vector<double>> M2(Matrix.begin() + w1, Matrix.end());
+    auto multiplyMatrixVectorPtr = static_cast<std::vector<double>(*)(const std::vector<std::vector<double>>&, const std::vector<double>&)>(&multiplyMatrixVector);
     Timer timer;
-    std::future<std::vector<float>> fu = std::async(std::launch::async, multiplyMatrixVectorPtr, std::cref(M1), std::cref(Vector));
-    std::vector<float> R2 = M2 * Vector;
-    std::vector<float> R1 = fu.get(); 
+    std::future<std::vector<double>> fu = std::async(std::launch::async, multiplyMatrixVectorPtr, std::cref(M1), std::cref(Vector));
+    std::vector<double> R2 = M2 * Vector;
+    std::vector<double> R1 = fu.get(); 
 
     R1.insert(R1.end(), R2.begin(), R2.end());
     return R1;
